Added autoloader, removed some overhead, added example
This version removes a lot of overhead by sending the data as
'application/json' back and forth rather than as 'text/html' or
'multipart/form-data' with boundaries.  This is also setup to interface
with a REST API.  I am already using this on a few of my website
backends to add, edit, delete, and load objects from a MySQL database
or text files.